fix: pin numeric precision to clear dbt1071 warnings
Every `dbt build` emitted 7 JinjaWarn (dbt1071) warnings: "columns with
numeric type and unspecified precision/scale". Fusion treats NUMBER(38,0) --
Snowflake's default for an unconstrained integer -- as unspecified, so an
explicit CAST to NUMBER(38,0) does not silence it; the precision has to be
genuinely narrower.

Casts are applied at the staging boundary so precision propagates downstream
rather than being patched per-mart:

  court_number      NUMBER(2,0)   stg_reservations, stg_lessons, stg_matches
  duration_minutes  NUMBER(6,0)   stg_lessons, stg_matches
  season_year       NUMBER(4,0)   stg_tournaments
  max_participants  NUMBER(6,0)   stg_tournaments
  seed              NUMBER(4,0)   stg_tournament_entries
  quantity          NUMBER(6,0)   stg_invoice_line_items, stg_sales_transactions
  stock_quantity    NUMBER(9,0)   stg_products

Derived columns in dim_members: age NUMBER(3,0), age_group NUMBER(3,0),
membership_length_in_months NUMBER(6,0).

Widths were chosen with headroom against observed maxima in RAW.ZTC (e.g.
duration_minutes max 178 -> NUMBER(6,0)); all ten columns are scale 0, so no
rounding behaviour changes. Verified no truncation: fct_matches still 415 rows
with max duration 178 and max court 5.

Knock-on changes:
- fct_reservations pinned court_number to NUMBER(38,0) (from 58c23bd); narrowed
  to NUMBER(2,0), which also resolves the mismatch in fct_reservation_events
  and fct_reservations_hwm since both inherit from it.
- gross_amount is quantity * unit_price, so narrowing quantity changes the
  result type from NUMBER(38,2) to NUMBER(14,2); contract updated to match.
- 25 core contract data_type entries updated from `number`/`NUMBER(38,0)` to
  the precise types.

member_id and duration_in_mins remain NUMBER(38,0) by intent (declared PK type
and an existing explicit pin); neither triggers dbt1071.

Verified with `dbt build --full-refresh --no-manage-state` (nothing reused, so
every model is genuinely re-evaluated): 207 success, 0 errors, 4 warnings --
all four the by-design string_relationships member-FK warns. sqlfluff,
pre-commit, and dbt-bouncer (221/221) all clean.

+++ models/staging/stg_reservations.sql
@@ -18,7 +18,7 @@ renamed AS (
             ELSE TO_TIME(eindtijd)
         END AS end_time,
         uren AS duration_in_hours,
-        CAST(RIGHT(banen, 1) AS INTEGER) AS court_number,
+        CAST(RIGHT(banen, 1) AS NUMBER(2, 0)) AS court_number,
         inschrijver_persoon1 AS player_1,
         persoon2 AS player_2,
         persoon3 AS player_3,
